+++ app/build.gradle.kts
@@ -28,6 +28,9 @@ android {
             )
         }
     }
+    testOptions {
+        unitTests.isReturnDefaultValues = true
+    }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
         targetCompatibility = JavaVersion.VERSION_11
@@ -59,4 +62,8 @@ dependencies {
     androidTestImplementation("io.mockk:mockk-android:1.13.5")
 
     implementation("androidx.fragment:fragment-ktx:1.5.7")
+
+    implementation("com.squareup.okhttp3:mockwebserver:4.11.0")
+    implementation("com.squareup.okhttp3:okhttp:4.11.0")
+    implementation("org.json:json:20210307")
 }
