+++ app/src/main/java/woowacourse/shopping/data/repository/remote/MockRemoteProductRepositoryImpl.kt
@@ -1,19 +1,18 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.repository.remote
 
 import com.example.domain.cache.ProductCache
 import com.example.domain.cache.ProductLocalCache
-import com.example.domain.datasource.productsDatasource
 import com.example.domain.model.Product
 import com.example.domain.repository.ProductRepository
+import woowacourse.shopping.data.service.MockProductRemoteService
 
-class ProductMockRepositoryImpl(
-    private val dataSource: List<Product> = productsDatasource,
+class MockRemoteProductRepositoryImpl(
+    private val service: MockProductRemoteService,
     override val cache: ProductCache = ProductLocalCache
 ) : ProductRepository {
-
     override fun getFirstProducts(): List<Product> {
         if (cache.productList.isEmpty()) {
-            val products = dataSource.take(LOAD_SIZE)
+            val products = service.request(0)
             cache.addProducts(products)
             return products
         }
@@ -21,19 +20,10 @@ class ProductMockRepositoryImpl(
     }
 
     override fun getNextProducts(lastProductId: Long): List<Product> {
-        val cacheData = cache.productList.filter { it.id > lastProductId }
-        if (cacheData.isNotEmpty()) return cacheData.take(LOAD_SIZE)
-
-        val newProducts = dataSource.filter { it.id > lastProductId }.take(LOAD_SIZE)
-        cache.addProducts(newProducts)
-        return newProducts
+        return service.request(lastProductId)
     }
 
     override fun resetCache() {
         cache.clear()
-    }
-
-    companion object {
-        private const val LOAD_SIZE = 20
     }
 }
