+++ app/src/main/java/woowacourse/shopping/feature/detail/dialog/CounterDialog.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.fragment.app.DialogFragment
 import woowacourse.shopping.common_ui.CounterView
-import woowacourse.shopping.data.CartRepositoryImpl
+import woowacourse.shopping.data.repository.local.CartRepositoryImpl
 import woowacourse.shopping.data.sql.cart.CartDao
 import woowacourse.shopping.databinding.DialogCounterBinding
 import woowacourse.shopping.model.ProductUiModel
