+++ domain/build.gradle
@@ -7,8 +7,3 @@ java {
     sourceCompatibility = JavaVersion.VERSION_11
     targetCompatibility = JavaVersion.VERSION_11
 }
-dependencies {
-    implementation("com.squareup.okhttp3:mockwebserver:4.11.0")
-    implementation("com.squareup.okhttp3:okhttp:4.11.0")
-    implementation("org.json:json:20210307")
-}
